+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # v0.4.0 (Jan 16, 2025)
 
 * Adapt home screen layout to use a table for simpler parsing of the information
+* Exchanged the total timer in the bottom for a dynamic timeline that shows a
+  miniature version of the progression alongside marks indicating at which
+  points reminders will be shown
+* The template wizard now provides a preview of the progression of various
+  reminders you add to give you a feeling for when reminders will be shown to
+  the presenters
 
 # v0.3.0 (Jan 15, 2025)